Update Product_DelimitedTable_VALID_lid_ref_PubAvail.xml
This file / test case needs to be deleted.  But, I couldn't figure out how to delete the file from the 'test' directory in this Build.

So, I modified the value for <software_availability> from the no longer valid value of "Publically Available" to one of the valid values.  I chose a redundant value of "Released".